csc/python/game-backend/requirements.txt
asn1crypto==0.24.0
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-heroku==0.3.1
django-rest-api==0.1.5
django-rest-auth==0.9.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
ndg-httpsclient==0.5.1
oauthlib==3.1.0
psycopg2==2.7.7
psycopg2-binary==2.8.4
pusher==2.1.4
pyasn1==0.4.7
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
